refacor: 전처리 로직 수정
- 가우시안 블러 제거, 모폴로지 연산 제거
- 테스트 코드 변경
- 테스트 파일 저장
